Add files via upload
Data Exploration is an open-ended process where we learn what our data can tell us. It generally starts out with a high level overview, then narrows in to specific areas as we find intriguing areas of the data. The findings may be interesting in their own right, or they can be used to inform our modeling choices, such as by helping us decide which features to use.
This contains EDA of Superstore Sales Data.